+++ core/fxcrt/css/cfx_cssstylerule.h
@@ -19,7 +19,7 @@ class CFX_CSSStyleRule {
   ~CFX_CSSStyleRule();
 
   size_t CountSelectorLists() const;
-  CFX_CSSSelector* GetSelectorList(int32_t index) const;
+  CFX_CSSSelector* GetSelectorList(size_t index) const;
   CFX_CSSDeclaration* GetDeclaration();
 
   void SetSelector(std::vector<std::unique_ptr<CFX_CSSSelector>>* list);
